refactor : 스터디룸 추천, 신청 완료 모달 컴포넌트 분리 및 swr 적용
