+++ components/Coretime/Gallery/index 2.ts
@@ -1,0 +1,1 @@
+export { default as GalleryModal } from './GalleryModal';
