+++ components/Search/Recommend/index.ts
@@ -1,0 +1,1 @@
+export { default as Recommend } from './Recommend';
