+++ hooks/FCM/index.ts
@@ -1,2 +1,3 @@
 export { default as useFCMPushAlarm } from './useFCMPushAlarm';
 export { default as useFirebaseInit } from './useFirebaseInit';
+export { default as ServiceWorkerManager } from './ServiceWorkerManager';
