+++ hooks/StudyRooms/index.ts
@@ -2,3 +2,5 @@ export { default as useGetMyStudyRoomList } from './useGetMyStudyRoomList';
 export { default as useGetLeaderStudyRoomList } from './useGetLeaderStudyRoomList';
 export { default as useGetApplyStudyRoomList } from './useGetApplyStudyRoomList';
 export { default as useGetRoomInfo } from './useGetRoomInfo';
+export { default as useGetSearchResult } from './useGetSearchResult';
+export { default as useGetRecommendRoomList } from './useGetRecommendRoomList';
