+++ pages/_app.tsx
@@ -8,34 +8,43 @@ import { RecoilRoot } from 'recoil';
 import { ThemeProvider } from 'styled-components';
 
 import { Header } from '@/components/Common/Header';
+import { ServiceWorkerManager } from '@/hooks/FCM';
 import { GlobalStyle } from '@/styles/GlobalStyle';
 import { theme } from '@/styles/theme';
 
 export default function App({ Component, pageProps }: AppProps) {
   const [isClient, setIsClient] = useState(false);
 
+  // const [captureTime, setIsCaptureTime] = useRecoilState(captureTimeState);
+
   useEffect(() => {
     setIsClient(true);
-    if ('serviceWorker' in navigator) {
-      window.addEventListener('load', () => {
-        navigator.serviceWorker.register('/sw.js').then(
-          (registration) => {
-            console.log('SW registered: ', registration);
-          },
-          (err) => {
-            console.log('SW registration failed: ', err);
-          },
-        );
-        navigator.serviceWorker.register('/firebase-messaging-sw.js').then(
-          (registration) => {
-            console.log('FCM registered: ', registration);
-          },
-          (err) => {
-            console.log('FCM registration failed: ', err);
-          },
-        );
-      });
-    }
+    // if ('serviceWorker' in navigator) {
+    //   window.addEventListener('load', () => {
+    //     navigator.serviceWorker.register('/sw.js').then(
+    //       (registration) => {
+    //         console.log('SW registered: ', registration);
+    //       },
+    //       (err) => {
+    //         console.log('SW registration failed: ', err);
+    //       },
+    //     );
+    //     navigator.serviceWorker.register('/firebase-messaging-sw.js').then(
+    //       (registration) => {
+    //         console.log('FCM registered: ', registration);
+    //       },
+    //       (err) => {
+    //         console.log('FCM registration failed: ', err);
+    //       },
+    //     );
+    //     navigator.serviceWorker.addEventListener('message', (event) => {
+    //       if (event.data.type === 'FCM_MESSAGE_RECEIVED') {
+    //         setIsCaptureTime((prev) => prev + 1);
+    //         console.log(captureTime);
+    //       }
+    //     });
+    //   });
+    // }
   }, []);
 
   if (!isClient) {
@@ -44,11 +53,13 @@ export default function App({ Component, pageProps }: AppProps) {
   if (typeof window === 'undefined') {
     return null;
   }
+
   return (
     isClient && (
       <RecoilRoot>
         <ThemeProvider theme={theme}>
           <GlobalStyle />
+          <ServiceWorkerManager />
           <Head>
             <title>LearniVerse</title>
             <link rel="icon" href="/favicon.ico" />
